% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\docType{package}
\name{biowomp}
\alias{-package}
\alias{biowomp}
\title{biowomp: Cluster-matching alluvial plots}
\description{
Main plotting function and helpers for bipartite-matching-based alluvial diagrams
}
